feat: add more size choices and reduce the default to 1KB
