+++ .github/workflows/playwright.yml
@@ -59,6 +59,6 @@ jobs:
       - uses: actions/upload-artifact@v4
         if: failure()
         with:
-          name: ${{ matrix.project }}-playwright-report
+          name: ${{ hashFiles(format('{0}/package.json', matrix.project)) }}-playwright-report
           path: ./${{ matrix.project }}/playwright-report/
           retention-days: 30
